pyproject.toml
[project]
name = "be-tcf"
version = "0.1.0"
authors = [
    { name = "Vadym Tatarinov", email = "utikpuhlik@protonmail.ch" }
]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "environs>=14.1.1",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "orjson>=3.10.15",
    "polars>=1.22.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "stackprinter>=0.2.12",
    "uvicorn>=0.34.0",
    "fastapi-pagination>=0.12.34",
    "prometheus-fastapi-instrumentator>=7.0.2",
    "sentry-sdk[fastapi]>=2.22.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.1",
    "elasticsearch[async]>=8.17.2",
    "pandas>=2.2.3",
    "gspread>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.38",
    "email-validator>=2.2.0",
    "pgvector>=0.4.0",
    "openai>=1.70.0",
    "tenacity>=9.1.2",
    "python-slugify>=8.0.4",
    "python-multipart>=0.0.20",
    "aioboto3>=14.3.0",
    "python-magic-bin>=0.4.14 ; sys_platform == 'win32'",
    "python-magic>=0.4.27 ; sys_platform != 'win32'",
    "celery>=5.5.2",
    "openpyxl>=3.1.5",
    "fastapi-users[oauth,sqlalchemy]>=14.0.1",
    "aiosmtplib>=4.0.1",
    "ddtrace>=3.9.1",
    "python-docx>=1.1.2",
    "testcontainers[postgres,redis]>=4.10.0",
]


[dependency-groups]
dev = [
    "coverage>=7.6.9",
    "pre-commit>=4.0.1",
    "psycopg2-binary>=2.9.10",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.4",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests", "app"]
asyncio_mode = "auto"
# https://github.com/pytest-dev/pytest-asyncio/issues/924
asyncio_default_fixture_loop_scope = "session"

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.ruff.lint]
extend-select = ["I", "C4", "SIM"]
ignore = []

[tool.ruff]
src = ["src"]
line-length = 88
indent-width = 4
show-fixes = true
target-version = "py312"
exclude = [
    ".git",
    ".git-rewrite",
    ".ipynb_checkpoints",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations"
]
